requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.1
defusedxml==0.7.1
Django==5.2.7
django-filter==25.1
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
drf-spectacular==0.28.0
drf-yasg==1.21.11
Faker==37.8.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
oauthlib==3.3.1
packaging==25.0
parameterized==0.9.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
